refactor: introduce shared package core
