+++ .dockerignore
@@ -1,0 +1,50 @@
+# Dependencies
+node_modules/
+**/node_modules/
+
+# Build outputs
+dist/
+**/dist/
+build/
+coverage/
+**/coverage/
+
+# Test and generated reports
+frontend/playwright-report/
+frontend/test-results/
+frontend/e2e/.auth/
+frontend/blob-report/
+
+# Runtime and local state
+data/
+doku/
+*.db
+*.sqlite
+*.sqlite3
+*.db-journal
+*.db-wal
+*.db-shm
+
+# Environment and logs
+.env
+.env.*
+!.env.example
+logs/
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+
+# Local workspace artifacts
+.git/
+.cache/
+.turbo/
+.specify/
+specs/
+
+# IDE and OS files
+.DS_Store
+Thumbs.db
+.idea/
+.vscode/settings.json
