+++ .github/workflows/test.yml
@@ -28,10 +28,12 @@ jobs:
           filters: |
             backend:
               - 'backend/**'
+              - 'shared/**'
               - 'biome.json'
               - '.github/workflows/test.yml'
             frontend:
               - 'frontend/**'
+              - 'shared/**'
               - 'biome.json'
               - '.github/workflows/test.yml'
 
@@ -58,7 +60,17 @@ jobs:
         with:
           node-version: '22'
           cache: 'npm'
-          cache-dependency-path: backend/package-lock.json
+          cache-dependency-path: |
+            shared/package-lock.json
+            backend/package-lock.json
+
+      - name: Install shared dependencies
+        working-directory: shared
+        run: npm ci
+
+      - name: Build shared core
+        working-directory: shared
+        run: npm run build
 
       - name: Install dependencies
         run: npm ci
@@ -106,7 +118,17 @@ jobs:
         with:
           node-version: '22'
           cache: 'npm'
-          cache-dependency-path: frontend/package-lock.json
+          cache-dependency-path: |
+            shared/package-lock.json
+            frontend/package-lock.json
+
+      - name: Install shared dependencies
+        working-directory: shared
+        run: npm ci
+
+      - name: Build shared core
+        working-directory: shared
+        run: npm run build
 
       - name: Install dependencies
         run: npm ci
