+++ backend/package.json
@@ -4,20 +4,25 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "NODE_ENV=development tsx watch src/index.ts",
-    "build": "tsc -p tsconfig.json",
+    "dev": "npm --prefix ../shared run build && NODE_ENV=development tsx watch src/index.ts",
+    "build": "npm --prefix ../shared run build && tsc -p tsconfig.json",
     "start": "node dist/index.js",
     "migrate": "tsx src/db/migrate.ts",
+    "pretest": "npm --prefix ../shared run build",
     "test": "vitest",
+    "pretest:run": "npm --prefix ../shared run build",
     "test:run": "vitest run",
+    "pretest:domain": "npm --prefix ../shared run build",
     "test:domain": "vitest run src/test/domain-safety.test.ts",
+    "pretest:coverage": "npm --prefix ../shared run build",
     "test:coverage": "vitest run --coverage",
     "lint": "npx biome check .",
     "lint:fix": "npx biome check --write .",
     "format": "npx biome format --write .",
-    "check": "npx biome check . && tsc --noEmit"
+    "check": "npm --prefix ../shared run build && npx biome check . && tsc --noEmit"
   },
   "dependencies": {
+    "@medassist/shared": "file:../shared",
     "@fastify/cookie": "^11.0.2",
     "@fastify/cors": "^11.2.0",
     "@fastify/formbody": "^8.0.2",
