+++ docker-compose.dev.yml
@@ -4,10 +4,12 @@ services:
   backend-dev:
     image: node:22-slim
     working_dir: /app
-    command: sh -c "chown -R node:node /app/node_modules /app/data && su node -c 'npm install && npm run dev'"
+    command: sh -c "chown -R node:node /app/node_modules /app/data /shared/node_modules && su node -c 'cd /shared && npm install && cd /app && npm install && npm run dev'"
     volumes:
       - ./backend:/app
+      - ./shared:/shared
       - backend_node_modules:/app/node_modules
+      - shared_node_modules:/shared/node_modules
       - ./data:/app/data
     env_file:
       - .env
@@ -29,10 +31,12 @@ services:
   frontend-dev:
     image: node:22-slim
     working_dir: /app
-    command: sh -c "chown -R node:node /app/node_modules && su node -c 'npm install && npm run dev -- --host --port 5173'"
+    command: sh -c "chown -R node:node /app/node_modules /shared/node_modules && su node -c 'cd /shared && npm install && cd /app && npm install && npm run dev -- --host --port 5173'"
     volumes:
       - ./frontend:/app
+      - ./shared:/shared
       - frontend_node_modules:/app/node_modules
+      - shared_node_modules:/shared/node_modules
     env_file:
       - .env
     environment:
@@ -43,8 +47,10 @@ services:
     security_opt:
       - no-new-privileges:true
     depends_on:
-      - backend-dev
+      backend-dev:
+        condition: service_healthy
 
 volumes:
   backend_node_modules:
   frontend_node_modules:
+  shared_node_modules:
