+++ frontend/package.json
@@ -4,18 +4,24 @@
   "version": "1.27.2",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
+    "dev": "npm --prefix ../shared run build && vite",
+    "build": "npm --prefix ../shared run build && vite build",
     "preview": "vite preview",
     "lint": "npx biome check .",
     "lint:fix": "npx biome check --write .",
     "format": "npx biome format --write .",
-    "check": "npx biome check . && tsc --noEmit",
+    "check": "npm --prefix ../shared run build && npx biome check . && tsc --noEmit",
+    "pretest": "npm --prefix ../shared run build",
     "test": "vitest",
+    "pretest:run": "npm --prefix ../shared run build",
     "test:run": "vitest run",
+    "pretest:coverage": "npm --prefix ../shared run build",
     "test:coverage": "vitest run --coverage",
+    "pretest:e2e": "npm --prefix ../shared run build",
     "test:e2e": "rm -rf test-results && playwright test --config=playwright.stable.config.ts",
+    "pretest:e2e:domain": "npm --prefix ../shared run build",
     "test:e2e:domain": "rm -rf test-results && playwright test --config=playwright.stable.config.ts e2e/domain-safety.spec.ts",
+    "pretest:e2e:all": "npm --prefix ../shared run build",
     "test:e2e:all": "rm -rf test-results && playwright test --config=playwright.all.config.ts",
     "test:e2e:local": "PLAYWRIGHT_HTML_OPEN=never PLAYWRIGHT_WORKERS=4 npm run test:e2e",
     "test:e2e:all:local": "PLAYWRIGHT_HTML_OPEN=never PLAYWRIGHT_WORKERS=4 npm run test:e2e:all",
@@ -28,6 +34,7 @@
     "test:e2e:report": "playwright show-report"
   },
   "dependencies": {
+    "@medassist/shared": "file:../shared",
     "i18next": "^26.2.0",
     "i18next-browser-languagedetector": "^8.2.1",
     "lucide-react": "^1.16.0",
