+++ package.json
@@ -3,10 +3,10 @@
   "private": true,
   "scripts": {
     "prepare": "husky",
-    "lint": "cd backend && npm run lint && cd ../frontend && npm run lint",
-    "lint:fix": "cd backend && npm run lint:fix && cd ../frontend && npm run lint:fix",
-    "check": "cd backend && npm run check && cd ../frontend && npm run check",
-    "build": "cd backend && npm run build && cd ../frontend && npm run build",
+    "lint": "cd shared && npm run check && cd ../backend && npm run lint && cd ../frontend && npm run lint",
+    "lint:fix": "cd shared && npx biome check --write . && cd ../backend && npm run lint:fix && cd ../frontend && npm run lint:fix",
+    "check": "cd shared && npm run check && cd ../backend && npm run check && cd ../frontend && npm run check",
+    "build": "cd shared && npm run build && cd ../backend && npm run build && cd ../frontend && npm run build",
     "test:domain": "cd backend && npm run test:domain && cd ../frontend && npm run test:e2e:domain"
   },
   "devDependencies": {
