+++ shared/src/timezone-regions.ts
@@ -1,0 +1,49 @@
+export const TIMEZONE_TO_REGION: Record<string, string> = {
+	"Europe/Berlin": "DE",
+	"Europe/Vienna": "AT",
+	"Europe/Zurich": "CH",
+	"Europe/London": "GB",
+	"Europe/Dublin": "IE",
+	"Europe/Paris": "FR",
+	"Europe/Madrid": "ES",
+	"Europe/Rome": "IT",
+	"Europe/Amsterdam": "NL",
+	"Europe/Brussels": "BE",
+	"Europe/Warsaw": "PL",
+	"Europe/Prague": "CZ",
+	"Europe/Stockholm": "SE",
+	"Europe/Oslo": "NO",
+	"Europe/Copenhagen": "DK",
+	"Europe/Helsinki": "FI",
+	"Europe/Athens": "GR",
+	"Europe/Lisbon": "PT",
+	"Europe/Moscow": "RU",
+	"Europe/Kiev": "UA",
+	"Europe/Kyiv": "UA",
+	"Europe/Budapest": "HU",
+	"Europe/Bucharest": "RO",
+	"America/New_York": "US",
+	"America/Chicago": "US",
+	"America/Denver": "US",
+	"America/Los_Angeles": "US",
+	"America/Phoenix": "US",
+	"America/Toronto": "CA",
+	"America/Vancouver": "CA",
+	"America/Mexico_City": "MX",
+	"America/Sao_Paulo": "BR",
+	"America/Buenos_Aires": "AR",
+	"Asia/Tokyo": "JP",
+	"Asia/Shanghai": "CN",
+	"Asia/Hong_Kong": "HK",
+	"Asia/Singapore": "SG",
+	"Asia/Seoul": "KR",
+	"Asia/Dubai": "AE",
+	"Asia/Kolkata": "IN",
+	"Australia/Sydney": "AU",
+	"Australia/Melbourne": "AU",
+	"Pacific/Auckland": "NZ",
+};
+
+export function getRegionForTimezone(timezone?: string | null): string | undefined {
+	return timezone ? TIMEZONE_TO_REGION[timezone] : undefined;
+}
